Detect when test methods might be missing
It doesn't make sense to have a group object with no tests attached to
it, so fail if we can't find any test methods.

If there's a difference in methods between your group object and
a pointer to that object, then we fail so we don't risk silently
omitting a test.

Signed-off-by: Mike Seplowitz <mseplowitz@bloomberg.net>