Deps: bump testify to switch to yaml.v3
NOTE: testgroup does not use gopkg.in/yaml.* directly.

gopkg.in/yaml.v2 has multiple vulnerabilities, and the project has moved
on to v3, so we're updating testgroup to depend on the earliest version
of github.com/stretchr/testify that uses gopkg.in/yaml.v3.

Users of testgroup should feel free use to use even newer versions of
testify and/or yaml.v3.

Signed-off-by: Mike Seplowitz <mseplowitz@bloomberg.net>